+++ src/Storage/Storage.Management.Test/ScenarioTests/StorageAccountTests.cs
@@ -232,5 +232,14 @@ namespace Microsoft.Azure.Commands.Management.Storage.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-NewSetAzureStorageAccountAllowSharedKeyAccess");
         }
+
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestNewAzureStorageAccountEnableNfsV3()
+        {
+            TestRunner.RunTestScript("Test-NewAzureStorageAccountEnableNfsV3");
+        }
+
     }
 }